+++ R/latex.R
@@ -77,7 +77,6 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
     sprintf("\\captionsetup{justification=%s,singlelinecheck=off}\n\\caption{%s}\n", cap_setup, cap)
   }
   lab <- if (is.na(lab <- make_label(ht))) "" else sprintf("\\label{%s}\n", lab)
-  if (nzchar(lab) && ! nzchar(cap)) warning("No caption set: LaTeX table labels may not work as expected.")
 
   pos_text <- switch(position(ht),
     wrapleft = ,
@@ -94,6 +93,7 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
           pos_text[1],
           resize_box[1],
           "\n\\begin{threeparttable}\n",
+          "\\setlength{\\tabcolsep}{0pt}\n",
           res,
           "\\end{threeparttable}\n",
           resize_box[2],
@@ -218,7 +218,7 @@ build_tabular <- function(ht) {
   ## CELL CONTENTS -------------
   ## inner_cell is empty except for the *bottom* left of a 'display area' (including 1x1)
   ## this avoids a problem with later cells overpainting borders etc.
-  ## - inner_cell has padding, alignment, wrap and row_height TeX added
+  ## inner_cell has padding, alignment, wrap and row_height TeX added
   ## inner_cell data comes from the 'display cell' at the top left of the display area
 
   inner_cell_bldc <- clean_contents(ht, type = "latex")[bl_dc]
@@ -262,20 +262,23 @@ build_tabular <- function(ht) {
   pad_bldc$bottom <- bottom_padding(ht)[bl_dc]
   align_bldc      <- real_align[bl_dc]
   valign_bldc     <- valign(ht)[bl_dc]
-  wrap_bldc       <- wrap(ht)[bl_dc]
+  wrap_bldc       <- wrap(ht)[bl_dc] & ! is.na(width(ht)) # tables without width turn wrapping off
 
   has_pad_bldc <- lapply(pad_bldc, Negate(is.na))
   pad_bldc <- lapply(pad_bldc, function (x) if (is.numeric(x)) sprintf("%.4gpt", x) else x)
   tpad_tex_bldc <- rep("", length(pad_bldc$top))
-  # tpad_tex_bldc[has_pad_bldc$top] <- sprintf("\\rule{0pt}{\\baselineskip+%s}",
-  #       pad_bldc$top[has_pad_bldc$top])
   tpad_tex_bldc[has_pad_bldc$top] <- sprintf("\\huxtpad{%s}", pad_bldc$top[has_pad_bldc$top])
   bpad_tex_bldc <- rep("", length(pad_bldc$bottom))
   bpad_vals_bldc <- pad_bldc$bottom[has_pad_bldc$bottom]
   bpad_tex_bldc[has_pad_bldc$bottom] <- sprintf("\\huxbpad{%s}", bpad_vals_bldc)
   align_tex_key <- c("left" = "\\raggedright ", "right" = "\\raggedleft ", "center" = "\\centering ")
   align_tex_bldc <- align_tex_key[align_bldc]
-  inner_cell_bldc <- paste0(tpad_tex_bldc, align_tex_bldc, inner_cell_bldc, bpad_tex_bldc)
+  lpad_tex_bldc  <- ifelse(has_pad_bldc$left & ! wrap_bldc,
+        sprintf("\\hspace{%s} ", pad_bldc$left), "")
+  rpad_tex_bldc  <- ifelse(has_pad_bldc$right & ! wrap_bldc,
+        sprintf(" \\hspace{%s}", pad_bldc$right), "")
+  inner_cell_bldc <- paste0(tpad_tex_bldc, align_tex_bldc, lpad_tex_bldc, inner_cell_bldc,
+        rpad_tex_bldc, bpad_tex_bldc)
 
   if (any(wrap_bldc)) {
     # reverse of what you think. "b" aligns the *bottom* of the text with the baseline
@@ -284,9 +287,11 @@ build_tabular <- function(ht) {
     valign_bldc <- valign_tex_key[valign_bldc]
     # XXX should be a way to speed up by only doing dc_idx cells. but we run again at some point...
     width_spec_bldc <- width_spec[bl_dc]
+    left_pad_bldc <- ifelse(has_pad_bldc$left, sprintf("\\hspace{%s}", pad_bldc$left), "")
     hpad_loss_left_bldc  <- ifelse(has_pad_bldc$left,  paste0("-", pad_bldc$left),  "")
     hpad_loss_right_bldc <- ifelse(has_pad_bldc$right, paste0("-", pad_bldc$right), "")
-    inner_cell_bldc[wrap_bldc] <- sprintf("\\parbox[%s]{%s%s%s}{%s}",
+    inner_cell_bldc[wrap_bldc] <- sprintf("%s\\parbox[%s]{%s%s%s}{%s}",
+            left_pad_bldc[wrap_bldc],
             valign_bldc[wrap_bldc],
             width_spec_bldc[wrap_bldc],
             hpad_loss_left_bldc[wrap_bldc],
@@ -335,7 +340,7 @@ build_tabular <- function(ht) {
   ## left borders are blank, except for the first row; we collapse borders into right border position
 
   colspan_lhdc    <- colspan(ht)[lh_dc]
-  wrap_lhdc       <- wrap(ht)[lh_dc]
+  wrap_lhdc       <- wrap(ht)[lh_dc] & ! is.na(width(ht))
   valign_lhdc     <- valign(ht)[lh_dc]
   real_align_lhdc <- real_align[lh_dc]
   colspec_tex_key <- c("left" = "l", "center" = "c", "right" = "r")
@@ -356,12 +361,10 @@ build_tabular <- function(ht) {
   bord <- cb$vert
   bcol <- cbc$vert
   has_bord <- ! is.na(bord)
-  has_bcol <- ! is.na(bcol) # if *defined* as black, then we print it. Otherwise not.
   bs_double <- cbs$vert == "double"
   bcol <- format_color(bcol, default = "black")
   bord_tex <- rep("", length(bord))
 
-  # bord_tex[has_bord] <- sprintf("!{%s\\vrule width %.4gpt}", bcol_tex[has_bord], bord[has_bord])
   bord_tex[has_bord] <- sprintf("!{\\huxvb{%s}{%.4g}}", bcol[has_bord], bord[has_bord])
   bord_tex[bs_double] <- paste0(bord_tex[bs_double], bord_tex[bs_double])
   dim(bord_tex) <- dim(cb$vert)
@@ -416,6 +419,7 @@ build_tabular <- function(ht) {
   table_body <- paste(hhlines[1], table_body, sep = "\n")
 
   tenv <- tabular_environment(ht)
+  if (is.na(tenv)) tenv <- if (is.na(width(ht))) "tabular" else "tabularx"
   tenv_tex <- paste0(c("\\begin{", "\\end{"), tenv, "}")
   width_spec <- if (tenv %in% c("tabularx", "tabular*", "tabulary")) {
     tw <- latex_table_width(ht)
@@ -424,9 +428,13 @@ build_tabular <- function(ht) {
     ""
   }
 
-  colspec_top <- sapply(seq_len(ncol(ht)), function (mycol) {
+  colspec_top <- if (is.na(width(ht))) {
+    rep("l", ncol(ht))
+  } else {
+    sapply(seq_len(ncol(ht)), function (mycol) {
           sprintf("p{%s}", compute_width(ht, mycol, mycol))
         })
+  }
   colspec_top <- paste0(colspec_top, collapse = " ")
   colspec_top <- sprintf("{%s}\n", colspec_top)
 
@@ -443,7 +451,7 @@ latex_table_width <- function (ht) {
 
 
 compute_width <- function (ht, start_col, end_col) {
-  table_width <- width(ht) # always defined, default is 0.5 (of \\textwidth)
+  table_width <- width(ht)
   if (is.numeric(table_width)) {
     table_unit  <- default_table_width_unit
     table_width <- as.numeric(table_width)