+++ tests/testthat.R
@@ -4,3 +4,4 @@ library(huxtable)
 
 we_are_in_R_CMD_check <- TRUE
 test_check("huxtable")
+