+++ tests/testthat/test-print.R
@@ -2,26 +2,6 @@
 context("Printing to various formats")
 
 
-test_that("LaTeX output examples unchanged", {
-  test_ex_same("to_latex")
-})
-
-
-test_that("Markdown output examples unchanged", {
-  test_ex_same("to_md")
-})
-
-
-test_that("HTML output examples unchanged", {
-  test_ex_same("to_html")
-})
-
-
-test_that("Screen output examples unchanged", {
-  test_ex_same("to_screen")
-})
-
-
 test_that("to_screen gives warning with colour if crayon not installed", {
   ht <- hux(a = 1:2)
   with_mock(requireNamespace = function (...) FALSE, {