+++ tests/testthat/test-row-col-functions.R
@@ -2,11 +2,6 @@
 context("Row and column functions")
 
 
-test_that("Row/column function examples unchanged", {
-  test_ex_same("every")
-  test_ex_same("final")
-})
-
 test_that("final", {
   dfr <- data.frame(a = 1:5, b = 1:5, d = 1:5, e = 1:5)
   expect_equivalent(final(2)(dfr, 1), 4:5)
@@ -15,10 +10,10 @@ test_that("final", {
 })
 
 
-test_that("every(), evens and odds", {
+test_that("stripe(), evens and odds", {
   dfr <- data.frame(a = 1:7, b = 1:7, d = 1:7, e = 1:7)
-  expect_equivalent(every(3, from = 1)(dfr, 1), c(1, 4, 7))
-  expect_equivalent(every(3, from = 1)(dfr, 2), c(1, 4))
+  expect_equivalent(stripe(3, from = 1)(dfr, 1), c(1, 4, 7))
+  expect_equivalent(stripe(3, from = 1)(dfr, 2), c(1, 4))
   expect_equivalent(evens(dfr, 1), c(2, 4, 6))
   expect_equivalent(evens(dfr, 2), c(2, 4))
   expect_equivalent(odds(dfr, 1), c(1, 3, 5, 7))